Nome da classe alterado
